Split source files into directories
We added new source files recently and we will in near time. So
for better organization, I split the source files into directories.
