+++ config/paths.js
@@ -70,7 +70,7 @@ module.exports = {
 	appBuild: resolveApp('build'),
 	appPublic: resolveApp('public'),
 	appHtml: resolveApp('public/index.html'),
-	appIndexJs: resolveModule(resolveApp, 'src/index'),
+	appIndexJs: resolveModule(resolveApp, 'src/playground/index'),
 	appPackageJson: resolveApp('package.json'),
 	appSrc: resolveApp('src'),
 	appTsConfig: resolveApp('tsconfig.json'),
