+++ src/playground/index.js
@@ -3,7 +3,7 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import '../css/react-datetime.css';
+import '../../css/react-datetime.css';
 import App from './App';
 
 ReactDOM.render(<App />, document.getElementById('root'));
