+++ src/views/DaysView.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import ViewNavigation from './ViewNavigation';
+import ViewNavigation from '../parts/ViewNavigation';
 
 export default class DaysView extends React.Component {
 	static defaultProps = {
